% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_login.R
\name{check_login}
\alias{check_login}
\title{Check that the login request was successful or not}
\usage{
check_login(response)
}
\arguments{
\item{response}{(httr::response) RT API login response}
}
\value{
(logical) TRUE if login was succesful, errors out otherwise
}
\description{
Check that the login request was successful or not
}
