% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\name{construct_newline_pairs}
\alias{construct_newline_pairs}
\title{Construct a string for params suitable for passing into an RT request}
\usage{
construct_newline_pairs(params)
}
\arguments{
\item{params}{(list) One or more key value pairs}
}
\value{
(character)
}
\description{
RT's API, in a few cases, takes a body of key value pairs that are colon
separated and each key value pair is newline separated. Each pair is also
run through \code{\link{compact}} to remove \code{NULL} elements.
}
