% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\name{parse_rt_properties}
\alias{parse_rt_properties}
\title{Parse typical RT properties as contained in an RT response body}
\usage{
parse_rt_properties(body)
}
\arguments{
\item{body}{(character) Response body from an \code{rt_response}}
}
\value{
List of properties
}
\description{
The code gives a basic idea of the format but it's basically
newline-separated key-value pairs with a ': ' between them. e.g.,
}
\details{
id: queue/1
Name: General
}
