% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_create.R
\name{parse_ticket_create_body}
\alias{parse_ticket_create_body}
\title{Parse an RT ticket create response body and return the ticket ID}
\usage{
parse_ticket_create_body(body)
}
\arguments{
\item{body}{(character) The ticket create response body}
}
\value{
(numeric) The ticket ID
}
\description{
This function essential parses the text:
\code{"# Ticket 1 created."}
}
