% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_user_create.R
\name{parse_user_create_body}
\alias{parse_user_create_body}
\title{Parse the response body from a call to \code{\link{rt_user_create}}}
\usage{
parse_user_create_body(body)
}
\arguments{
\item{body}{(character)}
}
\value{
(numeric) The user ID
}
\description{
Parse the response body from a call to \code{\link{rt_user_create}}
}
