% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\name{print.rt_api}
\alias{print.rt_api}
\title{Print an \code{rt_api} object}
\usage{
\method{print}{rt_api}(x, max_lines = 10)
}
\arguments{
\item{x}{object of class \code{rt_api}}

\item{max_lines}{(numeric) The maximum number of lines of the body to print.}
}
\description{
Print an \code{rt_api} object
}
