% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_package.R
\docType{package}
\name{rt}
\alias{rt}
\title{The \code{rt} package}
\description{
\code{rt} provides a programming interface to the
\href{https://rt-wiki.bestpractical.com/wiki/REST}{Request Tracker API}.
}
\details{
Everything should be implemented and all functions should return a reasonably
useful result that's suitable for programming with.

See \href{https://github.com/NCEAS/rt}{GitHub} for more information.
}
