% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_login.R
\name{rt_login}
\alias{rt_login}
\title{Log in to RT}
\usage{
rt_login(base, user, pass)
}
\arguments{
\item{base}{(character) The base URL that hosts RT for your organization}

\item{user}{(character) Your username}

\item{pass}{(character) Your password}
}
\description{
Use this to log into RT at the start of your session.
}
\examples{
\dontrun{
rt_login("https://server.name/rt/", "my_username", "my_password")
}
}
