% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_queue_properties.R
\name{rt_queue_properties}
\alias{rt_queue_properties}
\title{Get the properties of a queue}
\usage{
rt_queue_properties(queue, ...)
}
\arguments{
\item{queue}{(character) The queue}

\item{...}{Other arguments passed to \code{\link{rt_GET}}}
}
\value{
(list) A list of queue properties
}
\description{
Get the properties of a queue
}
\examples{
\dontrun{
# By default, RT installations come with a General queue
# We can get its properties like this
rt_queue_properties("General")
}
}
