% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_search.R
\name{rt_search}
\alias{rt_search}
\title{Search RT}
\usage{
rt_search(query, orderBy = NULL, format = "l",
  rt_base = getOption("rt_base"))
}
\arguments{
\item{query}{(character) A query}

\item{orderBy}{(character) How to order your search results}

\item{format}{(character) Either \code{i} (ticket ID only),
\code{s} (ticket ID and subject), or \code{l} (full ticket metadata).
Defaults to \code{l}.}

\item{rt_base}{(character) The base URL that hosts RT for your organization. Set the base URL in your R session using \code{options(rt_base = "https://server.name/rt/")}}
}
\description{
Search RT for tickets that fit your query.
}
\examples{
\dontrun{
# To return all un-owned tickets on a queue:
rt_search(query = "Queue='some_queue'AND(Owner='Nobody')")
}
}
