% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket.R
\name{rt_ticket}
\alias{rt_ticket}
\title{Get ticket properties}
\usage{
rt_ticket(ticket_id, rt_base = getOption("rt_base"))
}
\arguments{
\item{ticket_id}{(numeric) The ticket identifier}

\item{rt_base}{(character) The base URL that hosts RT for your organization. Set the base URL in your R session using \code{options(rt_base = "https://server.name/rt/")}}
}
\description{
Retrieves ticket properties
}
\examples{
\dontrun{
rt_ticket_properties(15)
}
}
