% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_attachment_content.R
\name{rt_ticket_attachment_content}
\alias{rt_ticket_attachment_content}
\title{Get ticket attachment content}
\usage{
rt_ticket_attachment_content(ticket_id, attachment_id,
  rt_base = getOption("rt_base"))
}
\arguments{
\item{ticket_id}{(numeric) The ticket identifier}

\item{attachment_id}{(numeric) The attachment identifier}

\item{rt_base}{(character) The base URL that hosts RT for your organization. Set the base URL in your R session using \code{options(rt_base = "https://server.name/rt/")}}
}
\description{
Retrieves attachment content
}
\examples{
\dontrun{
rt_ticket_attachment_content(2, 1)
}
}
