% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_history_comment.R
\name{rt_ticket_history_comment}
\alias{rt_ticket_history_comment}
\title{Add ticket history comment}
\usage{
rt_ticket_history_comment(ticket_id, comment_text,
  rt_base = getOption("rt_base"))
}
\arguments{
\item{ticket_id}{(numeric) The ticket identifier}

\item{comment_text}{(character) Text that to add as a comment}

\item{rt_base}{(character) The base URL that hosts RT for your organization. Set the base URL in your R session using \code{options(rt_base = "https://server.name/rt/")}}
}
\description{
Add a comment to an existing ticket
}
\examples{
\dontrun{
rt_ticket_history_comment()
}
}
