% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_history_entry.R
\name{rt_ticket_history_entry}
\alias{rt_ticket_history_entry}
\title{Gets the history information for a single history item}
\usage{
rt_ticket_history_entry(ticket_id, history_id, ...)
}
\arguments{
\item{ticket_id}{(numeric) The ticket identifier}

\item{history_id}{(numeric) The history entry identifier}

\item{...}{Other arguments passed to \code{\link{rt_GET}}}
}
\value{
(rt_api) An \code{rt_api} object with the response
}
\description{
Gets the history information for a single history item
}
\examples{
\dontrun{
# Get the history entry for ticket 992 and history id 123
rt_ticket_history(992, 123)
}
}
