% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_history_reply.R
\name{rt_ticket_history_reply}
\alias{rt_ticket_history_reply}
\title{Add ticket history reply}
\usage{
rt_ticket_history_reply(ticket_id, text, cc = NULL, bcc = NULL,
  time_worked = "0", attachment_path = NULL, ...)
}
\arguments{
\item{ticket_id}{(numeric) The ticket identifier}

\item{text}{(character) Text that to add as a comment}

\item{cc}{(character) Email for cc}

\item{bcc}{(character) Email for bcc}

\item{time_worked}{(character)}

\item{attachment_path}{(character) Path to a file to upload}

\item{...}{Other arguments passed to \code{\link{rt_POST}}}
}
\description{
Add a response to an existing ticket
}
\examples{
\dontrun{
rt_ticket_history_reply(11,
                        "Thank you.

                        Have a great day!")
}
}
