% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_links.R
\name{rt_ticket_links}
\alias{rt_ticket_links}
\title{Get RT ticket links}
\usage{
rt_ticket_links(ticket_id, rt_base_url = Sys.getenv("RT_BASE_URL"))
}
\arguments{
\item{ticket_id}{(numeric) The ticket identifier}

\item{rt_base_url}{(character) The address of your RT installation.
See details.}
}
\description{
Gets the ticket links for a single ticket. If applicable, the following fields will be returned: \code{HasMember},
\code{ReferredToBy}, \code{DependedOnBy}, \code{MemberOf}, \code{RefersTo}, and \code{DependsOn}.
}
\examples{
\dontrun{
rt_ticket_links(1007)
}
}
