% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_merge.R
\name{rt_ticket_merge}
\alias{rt_ticket_merge}
\title{Merge two tickets}
\usage{
rt_ticket_merge(origin, into)
}
\arguments{
\item{origin}{(character|numeric) Ticket ID to merge into \code{into}}

\item{into}{(character|numeric) Ticket ID to merge \code{origin} into}
}
\value{
(numeric) The ID of ticket both tickets were merged into
}
\description{
Merge two tickets
}
\examples{
\dontrun{
# First, create two tickets
ticket_one <- rt_ticket_create("General")
ticket_two <- rt_ticket_create("General")

# Then merge them together
ticket_merge(ticket_one, ticket_two)
}
}
