% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_properties.R
\name{rt_ticket_properties}
\alias{rt_ticket_properties}
\title{Get ticket properties}
\usage{
rt_ticket_properties(ticket_id, rt_base_url = Sys.getenv("RT_BASE_URL"))
}
\arguments{
\item{ticket_id}{(numeric) The ticket identifier}

\item{rt_base_url}{(character) The address of your RT installation.
See details.}
}
\description{
Retrieves ticket properties
}
\examples{
\dontrun{
rt_ticket_properties(15)
}
}
