% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_user.R
\name{rt_user}
\alias{rt_user}
\title{Get RT user properties}
\usage{
rt_user(user_id, rt_base = getOption("rt_base"))
}
\arguments{
\item{user_id}{(numeric|character) The user id or login.}

\item{rt_base}{(character) The base URL that hosts RT for your organization. Set the base URL in your R session using \code{options(rt_base = "https://server.name/rt/")}}
}
\description{
Gets the ticket links for a single ticket. If applicable, the following fields will be returned: \code{HasMember},
\code{ReferredToBy}, \code{DependedOnBy}, \code{MemberOf}, \code{RefersTo}, and \code{DependsOn}.
}
\examples{
\dontrun{
rt_user_properties(1)
}
}
