% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\name{rt_user_agent}
\alias{rt_user_agent}
\title{Get the user agent for the package.}
\usage{
rt_user_agent()
}
\value{
(character) The user agent string for the package
}
\description{
This is used by \link{\code{rt_GET}} and \link{\code{rt_POST}} to provide
HTTP requests with an appropriate user agent.
}
\examples{
rt_user_agent()
}
