% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_user_properties.R
\name{rt_user_properties}
\alias{rt_user_properties}
\title{Get RT user properties}
\usage{
rt_user_properties(user_id, rt_base_url = Sys.getenv("RT_BASE_URL"))
}
\arguments{
\item{user_id}{(numeric) The user identifier}

\item{rt_base_url}{(character) The address of your RT installation.
See details.}
}
\description{
Gets the ticket links for a single ticket. If applicable, the following fields will be returned: \code{HasMember},
\code{ReferredToBy}, \code{DependedOnBy}, \code{MemberOf}, \code{RefersTo}, and \code{DependsOn}.
}
\examples{
\dontrun{
rt_user_properties(1)
}
}
