% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\name{rt_version_string}
\alias{rt_version_string}
\title{Get the version of the currently installed version of this package as a
character vector}
\usage{
rt_version_string()
}
\value{
(character) The version is a character vector, e.g. "1.2.3"
}
\description{
Get the version of the currently installed version of this package as a
character vector
}
\examples{
get_package_version_string()
}
