% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{stopforstatus}
\alias{stopforstatus}
\title{Throw an error if the RT status code is an error status}
\usage{
stopforstatus(response)
}
\arguments{
\item{response}{(response) An \code{httr} response object}
}
\value{
Either nothing, or throws an error
}
\description{
Throw an error if the RT status code is an error status
}
