% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_search.R
\name{tidy_long_search_result}
\alias{tidy_long_search_result}
\title{tidy_long_search_result}
\usage{
tidy_long_search_result(result)
}
\arguments{
\item{result}{(list) List of lists from search results}
}
\value{
A \code{data.frame} or \code{tibble}
}
\description{
tidy_long_search_result
}
