% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_search.R
\name{try_tibble}
\alias{try_tibble}
\title{Try to make a tibble}
\usage{
try_tibble(df, coerce = TRUE)
}
\arguments{
\item{df}{(data.frame) The \code{data.frame} to try attempt to coerce to a
\code{tibble}}

\item{coerce}{(logical) Whether or not to try coercion. Provided for upstream
calling functions.}
}
\value{
Either a \code{data.frame} or a \code{tibble}
}
\description{
Try to make a tibble
}
