% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\name{print.rt_api}
\alias{print.rt_api}
\title{Print an \code{rt_api} object}
\usage{
\method{print}{rt_api}(x, ...)
}
\arguments{
\item{x}{object of class \code{rt_api}}

\item{...}{Other arguments passed to \code{\link[utils]{head}}}
}
\description{
Print an \code{rt_api} object
}
