% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\name{rt_GET}
\alias{rt_GET}
\title{Get an RT response}
\usage{
rt_GET(url)
}
\arguments{
\item{url}{(character) The base URL that hosts RT for your organization}
}
\description{
Get an RT response and format it into an S3 object
}
