% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_login.R
\name{rt_login}
\alias{rt_login}
\title{Log in to RT}
\usage{
rt_login(user, pass, rt_base = getOption("rt_base"))
}
\arguments{
\item{user}{(character) Your username}

\item{pass}{(character) Your password}

\item{rt_base}{(character) The base URL that hosts RT for your organization. Set the base URL in your R session using \code{options(rt_base = "https://server.name/rt/")}}
}
\description{
Use this to log into RT at the start of your session.
}
\examples{
\dontrun{
options(rt_base = "https://demo.bestpractical.com/")
rt_login("my_username", "my_password")
}
}
