% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_login.R
\name{rt_login_interactive}
\alias{rt_login_interactive}
\title{Log in to RT interactively}
\usage{
rt_login_interactive(base)
}
\arguments{
\item{base}{(character) The base URL that hosts RT for your organization}
}
\description{
Wrapper for \code{\link{rt_login}} to interactively log into RT at the start of your
session. Keeps your log-in information private.
}
\examples{
\dontrun{
rt_login_interactive("https://server.name/rt/")
}
}
