% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\name{rt_POST}
\alias{rt_POST}
\title{POST an RT request}
\usage{
rt_POST(url, ...)
}
\arguments{
\item{url}{(character) The full RT URL}

\item{...}{Other arguments passed to \code{\link[httr]{POST}}}
}
\value{
(rt_api) The parsed response from RT
}
\description{
POST an RT request
}
