% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_queue_properties.R
\name{rt_queue_properties}
\alias{rt_queue_properties}
\title{Get queue properties}
\usage{
rt_queue_properties(queue_id, rt_base = getOption("rt_base"))
}
\arguments{
\item{queue_id}{(character) The queue identifier}

\item{rt_base}{(character) The base URL that hosts RT for your organization. Set the base URL in your R session using \code{options(rt_base = "https://server.name/rt/")}}
}
\description{
Get the data for a single queue.
}
\examples{
\dontrun{
rt_queue_properties("General")
}
}
