% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_queue_properties.R
\name{rt_queue_properties}
\alias{rt_queue_properties}
\title{Get queue properties}
\usage{
rt_queue_properties(queue)
}
\arguments{
\item{queue}{(character) The queue}
}
\description{
Get the data for a single queue.
}
\examples{
\dontrun{
rt_queue_properties("General")
}
}
