% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_queue_properties.R
\name{rt_queue_properties}
\alias{rt_queue_properties}
\title{Get queue properties}
\usage{
rt_queue_properties(queue, rt_base_url = Sys.getenv("RT_BASE_URL"))
}
\arguments{
\item{queue}{(character) The queue}

\item{rt_base_url}{(character) The address of your RT installation.
See details.}
}
\description{
Get the data for a single queue.
}
\examples{
\dontrun{
rt_queue_properties("General")
}
}
