% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_search.R
\name{rt_search}
\alias{rt_search}
\title{Search RT}
\usage{
rt_search(base, query, orderBy = NULL, format = "l")
}
\arguments{
\item{base}{(character) The base URL that hosts RT for your organization}

\item{query}{(character) A query}

\item{orderBy}{(character) How to order your search results}

\item{format}{(character) Either \code{i} (ticket ID only),
\code{s} (ticket ID and title), or \code{l} (full ticket metadata).
Defaults to \code{l}.}
}
\description{
Search RT for tickets that fit your query.
}
\examples{
\dontrun{
# To return all un-owned tickets on a queue:
rt_search("https://server.name/rt/",
query="Queue='some_queue'AND(Owner='Nobody')")
}
}
