% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_attachment.R
\name{rt_ticket_attachment}
\alias{rt_ticket_attachment}
\title{Get ticket attachment}
\usage{
rt_ticket_attachment(ticket_id, attachment_id,
  rt_base = getOption("rt_base"))
}
\arguments{
\item{ticket_id}{(numeric) The ticket identifier}

\item{attachment_id}{(numeric) The attachment identifier}

\item{rt_base}{(character) The base URL that hosts RT for your organization. Set the base URL in your R session using \code{options(rt_base = "https://server.name/rt/")}}
}
\description{
Retrieves attachment metadata
}
\examples{
\dontrun{
rt_ticket_attachment(2, 1)
}
}
\seealso{
Other attachments: \code{\link{rt_ticket_attachment_content}},
  \code{\link{rt_ticket_attachments}}
}
\concept{attachments}
