% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_attachment.R
\name{rt_ticket_attachment}
\alias{rt_ticket_attachment}
\title{Get a ticket's attachment}
\usage{
rt_ticket_attachment(ticket_id, attachment_id, ...)
}
\arguments{
\item{ticket_id}{(numeric) The ticket identifier}

\item{attachment_id}{(numeric) The attachment identifier}

\item{...}{Other arguments passed to \code{\link{rt_GET}}}
}
\description{
Retrieves attachment metadata. To get the attachment itself, see
\link{rt_ticket_attachment_content}.
}
\examples{
\dontrun{
rt_ticket_attachment(2, 1)
}
}
