% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_attachment.R
\name{rt_ticket_attachment}
\alias{rt_ticket_attachment}
\title{Get ticket attachment}
\usage{
rt_ticket_attachment(base, ticket, attachment)
}
\arguments{
\item{base}{(character) The base URL that hosts RT for your organization}

\item{ticket}{(numeric) The ticket identifier}

\item{attachment}{(numeric) The attachment identifier}
}
\description{
Retrieves attachment information using the GET method.
}
\examples{
\dontrun{
rt_ticket_attachment("https://server.name/rt/", 12345, 56789)
}
}
