% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_attachment_content.R
\name{rt_ticket_attachment_content}
\alias{rt_ticket_attachment_content}
\title{Get ticket attachment content}
\usage{
rt_ticket_attachment_content(ticket_id, attachment_id,
  rt_base_url = Sys.getenv("RT_BASE_URL"))
}
\arguments{
\item{ticket_id}{(numeric) The ticket identifier}

\item{attachment_id}{(numeric) The attachment identifier}

\item{rt_base_url}{(character) The address of your RT installation.
See details.}
}
\description{
Retrieves attachment content
}
\examples{
\dontrun{
rt_ticket_attachment_content(2, 1)
}
}
