% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_attachment_content.R
\name{rt_ticket_attachment_content}
\alias{rt_ticket_attachment_content}
\title{Get the content of an attachment}
\usage{
rt_ticket_attachment_content(ticket_id, attachment_id, ...)
}
\arguments{
\item{ticket_id}{(numeric) The ticket identifier}

\item{attachment_id}{(numeric) The attachment identifier}

\item{...}{Other arguments passed to \code{\link{rt_GET}}}
}
\description{
Get the content of an attachment
}
\examples{
\dontrun{
rt_ticket_attachment_content(2, 1)
}
}
