% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_attachments.R
\name{rt_ticket_attachments}
\alias{rt_ticket_attachments}
\title{Get ticket attachments}
\usage{
rt_ticket_attachments(ticket_id, rt_base = getOption("rt_base"))
}
\arguments{
\item{ticket_id}{(numeric) The ticket identifier}

\item{rt_base}{(character) The base URL that hosts RT for your organization. Set the base URL in your R session using \code{options(rt_base = "https://server.name/rt/")}}
}
\description{
Retrieves attachment metadata for a ticket
}
\examples{
\dontrun{
rt_ticket_attachments(2)
}
}
\seealso{
Other attachments: \code{\link{rt_ticket_attachment_content}},
  \code{\link{rt_ticket_attachment}}
}
\concept{attachments}
