% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_attachments.R
\name{rt_ticket_attachments}
\alias{rt_ticket_attachments}
\title{Get a ticket's attachments}
\usage{
rt_ticket_attachments(ticket_id, ...)
}
\arguments{
\item{ticket_id}{(numeric) The ticket identifier}

\item{...}{Other arguments passed to \code{\link{rt_POST}}}
}
\value{
Either a \code{data.frame} or \code{tibble} of the attachments.
}
\description{
Retrieves attachment metadata for a ticket in a tabular form.
}
\examples{
\dontrun{
# Given a ticket exists with id '2', we can get its attachments as a table
rt_ticket_attachments(2)
}
}
