% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_history.R
\name{rt_ticket_history}
\alias{rt_ticket_history}
\title{Get ticket history}
\usage{
rt_ticket_history(base, ticket)
}
\arguments{
\item{base}{(character) The base URL that hosts RT for your organization}

\item{ticket}{(numeric) The ticket identifier}
}
\description{
Retrieves attachment information using the GET method.
}
\examples{
\dontrun{
rt_ticket_history("https://server.name/rt/", 12345)
}
}
