% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_history_comment.R
\name{rt_ticket_history_comment}
\alias{rt_ticket_history_comment}
\title{Add ticket history comment}
\usage{
rt_ticket_history_comment(ticket_id, comment_text, ...)
}
\arguments{
\item{ticket_id}{(numeric) The ticket identifier}

\item{comment_text}{(character) Text that to add as a comment}

\item{...}{Other arguments passed to \code{\link{rt_POST}}}
}
\description{
Add a comment to an existing ticket
}
\examples{
\dontrun{
rt_ticket_history_comment(1, "Your comment here...")
}
TODO: implement attachment
https://rt-wiki.bestpractical.com/wiki/REST#Ticket_History_Comment
}
