% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_history_comment.R
\name{rt_ticket_history_comment}
\alias{rt_ticket_history_comment}
\title{Add ticket history comment}
\usage{
rt_ticket_history_comment(ticket_id, comment_text,
  rt_base_url = Sys.getenv("RT_BASE_URL"))
}
\arguments{
\item{ticket_id}{(numeric) The ticket identifier}

\item{comment_text}{(character) Text that to add as a comment}

\item{rt_base_url}{(character) The address of your RT installation.
See details.}
}
\description{
Add a comment to an existing ticket
}
\examples{
\dontrun{
rt_ticket_history_comment()
}
}
