% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_links.R
\name{rt_ticket_links}
\alias{rt_ticket_links}
\title{Get RT ticket links}
\usage{
rt_ticket_links(ticket_id, rt_base = getOption("rt_base"))
}
\arguments{
\item{ticket_id}{(numeric|character) The ticket number}

\item{rt_base}{(character) The base URL that hosts RT for your organization. Set the base URL in your R session using \code{options(rt_base = "https://server.name/rt/")}}
}
\description{
Gets the ticket links for a single ticket. If applicable, the following fields will be returned: \code{HasMember},
\code{ReferredToBy}, \code{DependedOnBy}, \code{MemberOf}, \code{RefersTo}, and \code{DependsOn}.
}
\examples{
\dontrun{
rt_ticket_links(1007)
}
}
