% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_links.R
\name{rt_ticket_links}
\alias{rt_ticket_links}
\title{Get a ticket's links}
\usage{
rt_ticket_links(ticket_id, ...)
}
\arguments{
\item{ticket_id}{(numeric) The ticket identifier}

\item{...}{Other arguments passed to \code{\link{rt_GET}}}
}
\value{
(rt_api) An \code{rt_api} object with the response
}
\description{
Gets the ticket links for a single ticket. If applicable, the following
fields will be returned: \code{HasMember},
\code{ReferredToBy}, \code{DependedOnBy}, \code{MemberOf}, \code{RefersTo},
and \code{DependsOn}.
}
\examples{
\dontrun{
# Assuming have a ticket with id 1007, we can get it links by calling
rt_ticket_links(1007)
}
}
