% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_links_edit.R
\name{rt_ticket_links_edit}
\alias{rt_ticket_links_edit}
\title{Edit the links on a ticket}
\usage{
rt_ticket_links_edit(
  ticket_id,
  referred_to_by = NULL,
  depended_on_by = NULL,
  member_of = NULL,
  refers_to = NULL,
  depends_on = NULL,
  ...
)
}
\arguments{
\item{ticket_id}{(numeric) The ticket identifier}

\item{referred_to_by}{Tickets that are referred to}

\item{depended_on_by}{Tickets that are depended on}

\item{member_of}{Ticket groups?}

\item{refers_to}{Tickets that are referred to}

\item{depends_on}{Tickets that are depended on}

\item{...}{Other arguments passed to \code{\link{rt_POST}}}
}
\value{
(numeric) The ID of the ticket
}
\description{
Edit the links on a ticket
}
\examples{
\dontrun{
# Make ticket 20 depend on ticket 21
rt_ticket_links_edit(20, depends_on = 21)
}

}
