% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_properties.R
\name{rt_ticket_properties}
\alias{rt_ticket_properties}
\title{Get ticket attachments}
\usage{
rt_ticket_properties(base, ticket)
}
\arguments{
\item{base}{(character) The base URL that hosts RT for your organization}

\item{ticket}{(numeric) The ticket identifier}
}
\description{
Retrieves attachment metadata using the GET method.
}
\examples{
\dontrun{
rt_ticket_properties("https://server.name/rt/", 12345)
}
}
