% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_properties.R
\name{rt_ticket_properties}
\alias{rt_ticket_properties}
\title{Get a ticket's properties}
\usage{
rt_ticket_properties(ticket_id, ...)
}
\arguments{
\item{ticket_id}{(numeric) The ticket identifier}

\item{...}{Other arguments passed to \code{\link{rt_GET}}}
}
\value{
(list) A list of the ticket's properties
}
\description{
Retrieves ticket properties
}
\examples{
\dontrun{
rt_ticket_properties(15)
}
}
