% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_search.R
\name{rt_ticket_search}
\alias{rt_ticket_search}
\title{Search RT}
\usage{
rt_ticket_search(
  query,
  orderby = NULL,
  format = "l",
  fields = NULL,
  coerce_tibble = TRUE,
  ...
)
}
\arguments{
\item{query}{(character) Your query (See Details)}

\item{orderby}{(character) How to order your search results. Should be a
ticket property name preceeded by either a + or a - character.}

\item{format}{(character) Either \code{i} (ticket ID only),
\code{s} (ticket ID and subject), or \code{l} (full ticket metadata).
Defaults to \code{l}.}

\item{fields}{(character) Comma-separated list of fields to include in the
results.}

\item{coerce_tibble}{(logical) Whether or not to make the result when
\code{format} is \code{l} or \code{s} into a \code{tibble}. Defaults to \code{TRUE} and only works
when the \code{tibble} package is installed.}

\item{...}{Other arguments passed to \code{\link{rt_GET}}}
}
\value{
Either a data.frame/tibble (when format is \code{l} or \code{s}) or a numeric
vector when its \code{i}.
}
\description{
Search RT for tickets using RT's query syntax which is documented at
\url{https://docs.bestpractical.com/rt/4.4.3/query_builder.html}.
}
\examples{
\dontrun{
# To return all un-owned tickets on a queue:
rt_ticket_search("Queue='General' AND (Status='new')")

# We can sort by date created, increasing
rt_ticket_search("Queue='General' AND (Status='new')",
                 orderby = "+Created")

# If we just need a vector of ticket ids
rt_ticket_search("Queue='General' AND (Status='new')",
                 orderby = "+Created",
                 format = "i")
}
}
