% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\name{rt_url}
\alias{rt_url}
\title{Generate an RT API URL}
\usage{
rt_url(..., base_url = Sys.getenv("RT_BASE_URL"))
}
\arguments{
\item{...}{Parts of the URL to be joined by "/"}

\item{base_url}{(character) The base URL that hosts RT for your organization}
}
\description{
Create an RT API URL based on the server URL and any arguments provided
}
