% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_user_create.R
\name{rt_user_create}
\alias{rt_user_create}
\title{Create new user}
\usage{
rt_user_create(user_id, password, name = NULL, email_address = NULL,
  real_name = NULL, organization = NULL, privileged = NULL,
  disabled = NULL, rt_base = getOption("rt_base"))
}
\arguments{
\item{user_id}{(numeric) The user identifier}

\item{password}{(character) The password}

\item{name}{(character) Optional. User name}

\item{email_address}{(character) Optional. User email}

\item{real_name}{(character) Optional. User real name}

\item{organization}{(character) Optional. User organization}

\item{privileged}{(numeric) Optional. User privilege status}

\item{disabled}{(numeric) Optional. User disabled status}

\item{rt_base}{(character) The base URL that hosts RT for your organization. Set the base URL in your R session using \code{options(rt_base = "https://server.name/rt/")}}
}
\description{
Add a new RT user
}
\examples{
\dontrun{
rt_user_create()
}
}
