% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_user_create.R
\name{rt_user_create}
\alias{rt_user_create}
\title{Create new user}
\usage{
rt_user_create(name = NULL, password = NULL, email_address = NULL,
  real_name = NULL, organization = NULL, privileged = NULL,
  disabled = NULL)
}
\arguments{
\item{name}{(character) Optional. User name}

\item{password}{(character) The password}

\item{email_address}{(character) Optional. User email}

\item{real_name}{(character) Optional. User real name}

\item{organization}{(character) Optional. User organization}

\item{privileged}{(numeric) Optional. User privilege status}

\item{disabled}{(numeric) Optional. User disabled status}
}
\description{
Create a new RT user
}
\examples{
\dontrun{
rt_user_create()
}
}
