% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_user_properties.R
\name{rt_user_properties}
\alias{rt_user_properties}
\title{Get RT user properties}
\usage{
rt_user_properties(user_id)
}
\arguments{
\item{user_id}{(numeric) The ID of the User to edit}
}
\description{
Get RT user properties
}
\examples{
\dontrun{
rt_user_properties(1)
}
}
