% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ticket_search.R
\name{tidy_long_search_result}
\alias{tidy_long_search_result}
\title{tidy_long_search_result}
\usage{
tidy_long_search_result(result, coerce_tibble = TRUE)
}
\arguments{
\item{result}{(list) List of lists from search results}

\item{coerce_tibble}{(logical) Whether or not to try to coerce the result to
a \code{tibble}.}
}
\value{
A \code{data.frame} or \code{tbl_df}
}
\description{
tidy_long_search_result
}
